+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Fix Slideshow animations
 - Fix Sortable component sorting animation when sorting between lists
 - Fix event registration if `el` option is array
+- Slideshow navigation items may now be `data-` prefixed too
 
 ## 3.0.0 beta 32 (October 27, 2017)